Lookup for text would not show after clicking the page and using key commands to launch. https://bugs.webkit.org/show_bug.cgi?id=232766 rdar://74562619
Reviewed by Wenson Hsieh.

When triggering the lookup command with command-control-D or the triple finger tap on mac, if you have previously
clicked the page, the lookup would time out and the popover would never show. This was because after clicking on the page,
an invisible caret selection would exist, which would change the path taken for highlighting and temporarily selecting text to
be sent to data detectors.
The solution was to set ignoreSelectionChanges on the temporary selection, and also to stop propagating temporary selection changes to accessibility, as
that should not be necessary and that is the code that cause a sync message (RootViewToScreen) to be sent to the UI process while it was sync awaiting our
reply to the sync message (DidPerformImmediateActionHitTest) sent earlier for the original request for data detectors.

* editing/FrameSelection.cpp:
(WebCore::FrameSelection::updateAndRevealSelection):
* page/TextIndicator.cpp:
(WebCore::TextIndicator::createWithRange):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@285365 268f45cc-cd09-0410-ab3c-d52691b4dbfc