[LFC][IFC] The first run in visual order may not generate a display box https://bugs.webkit.org/show_bug.cgi?id=232806
Reviewed by Antti Koivisto.

Do not rely on the 0 index value to compute the content start position when visual ordering is in place.
The very first run (i == 0) in visual order may not even call into the displayBoxRect function (e.g. inline box end as </span>)
which means that we end up with an incorrect initial content start position (as we miss the RTL conversion).

* layout/formattingContexts/inline/InlineDisplayContentBuilder.cpp:
(WebCore::Layout::InlineDisplayContentBuilder::createBoxesAndUpdateGeometryForLineContent):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@285403 268f45cc-cd09-0410-ab3c-d52691b4dbfc