[iOS] Add a position information bit to indicate whether the hit-tested element is a paused video https://bugs.webkit.org/show_bug.cgi?id=232861
Reviewed by Megan Gardner.

Add `InteractionInformationAtPosition::isPausedVideo`, a flag that is true when the position information request
is over a paused video element (or inside the media control shadow root underneath that paused video element, in
the case where native controls are shown).

* Shared/ios/InteractionInformationAtPosition.h:
* Shared/ios/InteractionInformationAtPosition.mm:
(WebKit::InteractionInformationAtPosition::encode const):
(WebKit::InteractionInformationAtPosition::decode):

Also rename `imageElementContext` to `hostImageOrVideoElementContext` to clarify that it (1) may now include
element contexts for video elements, and (2) unlike the regular `elementContext`, this includes the image or
video element that is the host for hit-tested content in the UA shadow root of the element corresponding to
`elementContext`.

* UIProcess/ios/WKContentViewInteraction.mm:
(-[WKContentView hasSelectablePositionAtPoint:]):
(-[WKContentView textInteractionGesture:shouldBeginAtPoint:]):
(-[WKContentView imageAnalysisGestureDidBegin:]):
* WebProcess/WebPage/ios/WebPageIOS.mm:
(WebKit::videoPositionInformation):

Additionally populate the `image` of the position information, in the case where `includeImageData` is set on
the incoming request.

(WebKit::hostVideoElementIgnoringImageOverlay):
(WebKit::imagePositionInformation):
(WebKit::elementPositionInformation):
(WebKit::WebPage::positionInformation):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@285500 268f45cc-cd09-0410-ab3c-d52691b4dbfc