REGRESSION(r272201): Safari showed red distortion on webview after using Web Inspector, returning from another app https://bugs.webkit.org/show_bug.cgi?id=231358 <rdar://problem/81505208>
Reviewed by Devin Rousso.

Source/WebCore:

Test: inspector/page/setShowPaintRects.html

When we added support for animating individual transform properties, we moved to a model where all animations
for a given CSS property were wrapped in a dedicated CAAnimationGroup when running accelerated. Those groups
have an infinite duration and thus would never call the -animationDidStop:finished: CAAnimationDelegate method.

An option would have been to set the delegate on all animations contained in a CAAnimationGroup in
PlatformCALayerCocoa::addAnimationForKey(), but as it turns out, the CAAnimationDelegate methods for children
of an animation group aren't fired.

Since the only current use for the -animationDidStop:finished: delegate method is to eventually message back
into WebInspectorClient::animationEndedForLayer() for opacity animations, which don't need to run contained in
groups, our approach to address the issue is to only group animations for transform-related properties and
leave other animations as simple leaf animations.

To do this, inside of GraphicsLayerCA::updateAnimations(), we replace the addAnimation() lambda with a new
addLeafAnimation() lambda which we use for opacity, background-color and filter animations.

In order to be able to test the successful removal of repaint rects as a result of the CAAnimationDelegate
method firing, Patrick Angle contributed the required changes under inspector/ to expose the number of paint
rects to layout tests.

* inspector/InspectorClient.h:
(WebCore::InspectorClient::paintRectCount const):
* inspector/InspectorController.cpp:
(WebCore::InspectorController::paintRectCount const):
* inspector/InspectorController.h:
* inspector/InspectorOverlay.h:
(WebCore::InspectorOverlay::paintRectCount const):
* platform/graphics/ca/GraphicsLayerCA.cpp:
(WebCore::GraphicsLayerCA::updateAnimations):
* testing/Internals.cpp:
(WebCore::Internals::inspectorHighlightRects):
(WebCore::Internals::inspectorPaintRectCount):
* testing/Internals.h:
* testing/Internals.idl:

Source/WebKit:

* WebProcess/Inspector/WebInspectorClient.h:

LayoutTests:

Add a new test, written by Patrick Angle, that tracks whether paint rects have appeared
and then disappeared as content repaints in an inspected page.

* inspector/page/setShowPaintRects-expected.txt: Added.
* inspector/page/setShowPaintRects.html: Added.
* platform/ios-wk1/TestExpectations:
* platform/mac-wk1/TestExpectations:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@285529 268f45cc-cd09-0410-ab3c-d52691b4dbfc