[WebXR] three.js demos don't work https://bugs.webkit.org/show_bug.cgi?id=232798 rdar://83559881
Reviewed by Myles C. Maxfield.

Any content using three.js for WebXR was failing for
a couple of reasons. Firstly, we were not correctly
restoring the framebuffer, read, draw and texture
bindings in the native code, and three.js was correctly
assuming it didn't need to rebind.

Secondly, we were not resolving the multisample
framebuffer if the context was created with no alpha.
The issue is that our framebuffer from XR always has
an alpha channel, and you can't blit from RGB to RGBA.

* Modules/webxr/WebXROpaqueFramebuffer.cpp:
(WebCore::WebXROpaqueFramebuffer::startFrame): Make sure to restore the
bound texture target when exiting the function. Also ensure that we
tell the resolved FBO where its texture data comes from.
(WebCore::WebXROpaqueFramebuffer::endFrame): Restore the framebuffer binding.
(WebCore::WebXROpaqueFramebuffer::setupFramebuffer): Use the correct format.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@285544 268f45cc-cd09-0410-ab3c-d52691b4dbfc