[GTK] Layout Test media/video-seek-with-negative-playback.html timeouts on the release bot. https://bugs.webkit.org/show_bug.cgi?id=135086
Reviewed by Xabier Rodriguez-Calvar.

Source/WebCore:

In some specific cases, an EOS GstEvent can happen right before a seek. The event is translated
by playbin as an EOS GstMessage and posted to the bus, waiting to be forwarded to the main thread.
The EOS message (now irrelevant after the seek) is received and processed right after the seek,
causing the termination of the media at the player private and upper levels. This can even happen
after the seek has completed (m_isSeeking already false).

This patch detects that condition by ensuring that the playback is coherent with the EOS message,
that is, if we're still playing somewhere inside the playable ranges, there should be no EOS at
all. If that's the case, it's considered to be one of those spureous EOS and is ignored.

Live streams (infinite duration) are special and we still have to detect legitimate EOS there, so
this message bailout isn't done in those cases.

Also refactored the code that queries the position to the sinks.

* platform/graphics/gstreamer/MediaPlayerPrivateGStreamer.cpp: Ignore EOS message if the playback position is inside the playback limits when they're finite. Refactored sink position query code as gstreamerPositionFromSinks().
* platform/graphics/gstreamer/MediaPlayerPrivateGStreamer.h: Added gstreamerPositionFromSinks().

LayoutTests:

* platform/glib/TestExpectations: Unskipped test.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@285586 268f45cc-cd09-0410-ab3c-d52691b4dbfc