AX: Add missing AXID variant to AXPropertyValueVariant https://bugs.webkit.org/show_bug.cgi?id=233004
Patch by Tyler Wilcock <tyler_w@apple.com> on 2021-11-11
Reviewed by Chris Fleizach.

In AXIsolatedObject::setObjectProperty, we set property values to be
AXIDs. However, there is no AXID variant in AXPropertyValueVariant,
so none of these property lookups ever returned values.

This patch adds AXID as a variant to AXPropertyValueVariant, fixing
~15 text-only failures.

* accessibility/isolatedtree/AXIsolatedTree.h:
Add AXID as a variant to AXPropertyValueVariant.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@285686 268f45cc-cd09-0410-ab3c-d52691b4dbfc