Stop a momentum scroll animation when receiving a momentum end event https://bugs.webkit.org/show_bug.cgi?id=233236
Reviewed by Tim Horton.

Source/WebCore:

The stream of momentum events is terminated by an "end" event under two circumstances:
1. The gesture finished naturally without interruption.
2. The user tapped the trackpad with two fingers.

Unfortunately we can't tell these apart, so we always have to stop the animation.

Sadly this reveals a hole in the testing infrastructure. All the "changed" momentum
events in the sequence handed to UIHelper.mouseWheelSequence() are ignored while the
animation is running, but the "ended" event stops the animation. So any test without
a reasonable number of "changed" events will terminate its animation prematurely.
We can't fix this without new testing infrastructure (webkit.org/b/233234).

Test: fast/scrolling/mac/momentum-animator-end-event-stops.html

* platform/ScrollingEffectsController.h:
* platform/mac/ScrollingEffectsController.mm:
(WebCore::ScrollingEffectsController::handleWheelEvent):

LayoutTests:

New test, and skip one that breaks now because we need a fix for webkit.org/b/233234.

* fast/scrolling/mac/momentum-animator-end-event-stops-expected.txt: Added.
* fast/scrolling/mac/momentum-animator-end-event-stops.html: Added.
* platform/mac-wk2/TestExpectations:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@285917 268f45cc-cd09-0410-ab3c-d52691b4dbfc