Limit logical width over-constrained direction check https://bugs.webkit.org/show_bug.cgi?id=233189
Patch by Rob Buis <rbuis@igalia.com> on 2021-11-17
Reviewed by Sergio Villar Senin.

Source/WebCore:

In computePositionedLogicalWidthUsing the over-constrained case always checks for RTL
direction, but this should not be done if we are not dealing with left and right properties,
i.e. only when we are not orthogonal with the container.
Finally computeLogicalLeftPositionedOffset should not do the flipping logic in the over-constrained case.

* rendering/RenderBox.cpp:
(WebCore::computeLogicalLeftPositionedOffset):
(WebCore::RenderBox::computePositionedLogicalWidthUsing const):
(WebCore::RenderBox::computePositionedLogicalWidthReplaced const):

LayoutTests:

* TestExpectations: Removed 4 tests that are now passing.
* platform/ios/TestExpectations: Skipped several tests that show 1px diff in the result.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@285923 268f45cc-cd09-0410-ab3c-d52691b4dbfc